+++ src/ts/Ads/AdUnits/AbstractAdUnit.ts
@@ -1,5 +1,4 @@
-import { AdMobSignalFactory } from 'AdMob/Utilities/AdMobSignalFactory';
-import { AdUnitContainer, Orientation } from 'Ads/AdUnits/Containers/AdUnitContainer';
+import { AdUnitContainer, IAdUnit, Orientation } from 'Ads/AdUnits/Containers/AdUnitContainer';
 import { IAdsApi } from 'Ads/IAds';
 import { UserPrivacyManager } from 'Ads/Managers/UserPrivacyManager';
 import { OperativeEventManager } from 'Ads/Managers/OperativeEventManager';
@@ -21,6 +20,7 @@ import { DeviceInfo } from 'Core/Models/DeviceInfo';
 import { Observable0 } from 'Core/Utilities/Observable';
 import { IARApi } from 'AR/AR';
 import { IPurchasingApi } from 'Purchasing/IPurchasing';
+import { AbstractPrivacy } from 'Ads/Views/AbstractPrivacy';
 
 export interface IAdUnitParameters<T extends Campaign> {
     forceOrientation: Orientation;
@@ -42,9 +42,10 @@ export interface IAdUnitParameters<T extends Campaign> {
     privacyManager: UserPrivacyManager;
     programmaticTrackingService: ProgrammaticTrackingService;
     gameSessionId?: number;
+    privacy: AbstractPrivacy;
 }
 
-export abstract class AbstractAdUnit {
+export abstract class AbstractAdUnit implements IAdUnit {
 
     public static setAutoClose(value: boolean) {
         AbstractAdUnit._autoClose = value;