+++ src/ts/Display/AdUnits/DisplayInterstitialAdUnitParametersFactory.ts
@@ -16,14 +16,12 @@ export class DisplayInterstitialAdUnitParametersFactory extends AbstractAdUnitPa
     }
 
     protected createParameters(baseParams: IAdUnitParameters<DisplayInterstitialCampaign>) {
-        const privacy = this.createPrivacy(baseParams);
 
-        const view = new DisplayInterstitial(baseParams.platform, baseParams.core, baseParams.deviceInfo, baseParams.placement, baseParams.campaign, privacy, this.showGDPRBanner(baseParams));
+        const view = new DisplayInterstitial(baseParams.platform, baseParams.core, baseParams.deviceInfo, baseParams.placement, baseParams.campaign, baseParams.privacy, this.showGDPRBanner(baseParams));
         return  {
             ... baseParams,
             webPlayerContainer: this._webPlayerContainer,
-            view,
-            privacy
+            view
         };
     }
 }