+++ src/ts/MRAID/EventBridge/MRAIDEventAdapter.ts
@@ -97,4 +97,8 @@ export abstract class MRAIDEventAdapter implements IMRAIDAdapter {
     protected handleClose() {
         this._handler.onBridgeClose();
     }
+
+    protected handleSendStats(totalTime: number, playTime: number, frameCount: number) {
+        this._handler.onBridgeSendStats(totalTime, playTime, frameCount);
+    }
 }