+++ src/ts/VAST/AdUnits/VastAdUnitFactory.ts
@@ -10,7 +10,6 @@ import { VastVideoEventHandler } from 'VAST/EventHandlers/VastVideoEventHandler'
 import { IVideoEventHandlerParams } from 'Ads/EventHandlers/BaseVideoEventHandler';
 import { IObserver2, IObserver3 } from 'Core/Utilities/IObserver';
 import { StreamType } from 'Core/Constants/Android/StreamType';
-import { Privacy } from 'Ads/Views/Privacy';
 
 export class VastAdUnitFactory extends AbstractAdUnitFactory<VastCampaign, IVastAdUnitParameters> {
 
@@ -67,8 +66,6 @@ export class VastAdUnitFactory extends AbstractAdUnitFactory<VastCampaign, IVast
                 parameters.core.DeviceInfo.Ios!.onVolumeChanged.unsubscribe(onVolumeChangeObserverIOS);
             }
         });
-
-        Privacy.setupReportListener(parameters.privacy, vastAdUnit);
 
         return vastAdUnit;
     }