+++ src/ts/XPromo/AdUnits/XPromoAdUnitFactory.ts
@@ -1,7 +1,5 @@
 import { AbstractAdUnitFactory } from 'Ads/AdUnits/AbstractAdUnitFactory';
-import { IAdUnitParameters } from 'Ads/AdUnits/AbstractAdUnit';
 import { XPromoCampaign } from 'XPromo/Models/XPromoCampaign';
-import { XPromoEndScreen } from 'XPromo/Views/XPromoEndScreen';
 import { XPromoOverlayEventHandler } from 'XPromo/EventHandlers/XPromoOverlayEventHandler';
 import { XPromoEndScreenEventHandler } from 'XPromo/EventHandlers/XPromoEndScreenEventHandler';
 import { XPromoVideoEventHandler } from 'XPromo/EventHandlers/XPromoVideoEventHandler';
@@ -9,7 +7,6 @@ import { IVideoEventHandlerParams } from 'Ads/EventHandlers/BaseVideoEventHandle
 import { XPromoOperativeEventManager } from 'XPromo/Managers/XPromoOperativeEventManager';
 import { Platform } from 'Core/Constants/Platform';
 import { CustomFeatures } from 'Ads/Utilities/CustomFeatures';
-import { Privacy } from 'Ads/Views/Privacy';
 import { IXPromoAdUnitParameters, XPromoAdUnit } from 'XPromo/AdUnits/XPromoAdUnit';
 import { IStoreHandlerParameters } from 'Ads/EventHandlers/StoreHandlers/StoreHandler';
 import { StoreHandlerFactory } from 'Ads/EventHandlers/StoreHandlers/StoreHandlerFactory';
@@ -57,7 +54,6 @@ export class XPromoAdUnitFactory extends AbstractAdUnitFactory<XPromoCampaign, I
                 }
             });
         }
-        Privacy.setupReportListener(parameters.privacy, xPromoAdUnit);
 
         return xPromoAdUnit;
     }