+++ CHANGELOG.md
@@ -1,9 +1,36 @@
-# Pending
+# Merged
+
+* Disabling offending hybrid tests [#8002](https://github.com/Applifier/unity-ads-webview/pull/8002)
+
+# Tue Jan 29 18:19:32 UTC 2019
+
+* End screen improvements for the full network [#7890](https://github.com/Applifier/unity-ads-webview/pull/7890)
+* WebPlayer MRAID with Disabled AB group [#7938](https://github.com/Applifier/unity-ads-webview/pull/7938)
+* Refactor Privacy Code [#7934](https://github.com/Applifier/unity-ads-webview/pull/7934)
+* Fix iOS cached file URLs when loading backup campaigns [#7980](https://github.com/Applifier/unity-ads-webview/pull/7980)
+
+# Mon Jan 28 02:25:46 UTC 2019
+
+* Splitting MRAIDEventHandler to Playable and Programmatic [#7780](https://github.com/Applifier/unity-ads-webview/pull/7780)
+
+# Fri Jan 25 01:19:46 UTC 2019
+
+* Staging Scripts Fixes [#7912](https://github.com/Applifier/unity-ads-webview/pull/7912)
+* Promo adds greater UI flexibility [#7757](https://github.com/Applifier/unity-ads-webview/pull/7757)
+* VPAID in VAST ad should parse as VPAID ad [#7932](https://github.com/Applifier/unity-ads-webview/pull/7932)
+
+# Wed Jan 23 19:05:33 UTC 2019
+
+* New GDPR Consent shown before ad show (not enabled in backend yet) [#7753](https://github.com/Applifier/unity-ads-webview/pull/7753)
+
+# Wed Jan 23 02:00:42 UTC 2019
 
 * Remove backupcampaign no fill investigation [#7886](https://github.com/Applifier/unity-ads-webview/pull/7886)
 * Buffer swapping logic changes for ARView on Android [#7104](https://github.com/Applifier/unity-ads-webview/pull/7104)
 * Fix PREPARE_ERROR missing js_errors from 2.0.5 downwards
 * Add Bitmango game IDs to share Cheetah features [#7910](https://github.com/Applifier/unity-ads-webview/pull/7910)
+* Support relative urls in VAST strict parser [#7887](https://github.com/Applifier/unity-ads-webview/pull/7887)
+* Add Diagnostics for Missing Start Events [#7911](https://github.com/Applifier/unity-ads-webview/pull/7911)
 
 # Sat Jan 19 01:17:37 UTC 2019